💡 refactor: ライセンス表記のためコードを移動
